grapegram_client/pyproject.toml
[tool.poetry]
name = "grapegram"
version = "0.1.0"
description = ""
authors = ["Blacgolyb <andrejomelnickij@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
eel = "^0.16.0"
jinja2 = "^3.1.2"
requests = "^2.31.0"
pydantic = "^1.10.8"
websocket-client = "^1.5.2"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
